refactor: split BLE configuration service into connection, GPS and authentication services
Fixes #111, Fixes #112, Fixes #113
